diff: migrate xdiff to Rust implementation
