--- rust/diff/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "diff"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-similar = "2"
-rust_alloc = { path = "../../rust_alloc" }
-[lib]
-name = "diff"
-path = "src/lib.rs"
-crate-type = ["rlib"]
