+++ rust_diff/Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 crate-type = ["staticlib"]
 
 [dependencies]
-similar = "2"
+rust_xdiff = { path = "../rust_xdiff" }
 rust_linematch = { path = "../rust_linematch" }
