+++ rust_xdiff/Cargo.toml
@@ -5,8 +5,7 @@ edition = "2021"
 
 [dependencies]
 similar = "2"
-diff = { path = "../rust/diff" }
 
 [lib]
 name = "rust_xdiff"
-crate-type = ["staticlib", "rlib"]
+crate-type = ["rlib"]
