+++ rust_xdiff/src/lib.rs
@@ -1,117 +1,141 @@
-use std::ffi::{CStr, CString};
-use std::os::raw::{c_char, c_int, c_long, c_void};
-use similar::{Algorithm, TextDiff};
-use diff::{mmfile_t, mmbuffer_t, xpparam_t, xdemitcb_t, xdemitconf_t,
-    xdl_diff as rs_xdl_diff, xdiff_out_unified as rs_xdiff_out_unified,
-    xdiff_out_indices as rs_xdiff_out_indices};
+use core::ffi::{c_char, c_int, c_long, c_void};
+use similar::{ChangeTag, TextDiff};
 
-fn diff_internal(a: &str, b: &str) -> String {
-    TextDiff::configure()
-        .algorithm(Algorithm::Patience)
-        .diff_lines(a, b)
-        .unified_diff()
-        .to_string()
+#[repr(C)]
+pub struct mmfile_t {
+    pub ptr: *const c_char,
+    pub size: c_long,
 }
 
-#[no_mangle]
-pub extern "C" fn vim_xdiff_diff(a: *const c_char, b: *const c_char) -> *mut c_char {
-    if a.is_null() || b.is_null() {
-        return std::ptr::null_mut();
-    }
-    let a_str = unsafe { CStr::from_ptr(a) }.to_string_lossy().into_owned();
-    let b_str = unsafe { CStr::from_ptr(b) }.to_string_lossy().into_owned();
-    let result = diff_internal(&a_str, &b_str);
-    CString::new(result).unwrap().into_raw()
+#[repr(C)]
+pub struct mmbuffer_t {
+    pub ptr: *mut c_char,
+    pub size: c_long,
 }
 
-#[no_mangle]
-pub extern "C" fn vim_xdiff_free(ptr: *mut c_char) {
-    if !ptr.is_null() {
-        unsafe { drop(CString::from_raw(ptr)); }
-    }
+#[repr(C)]
+pub struct xpparam_t {
+    pub flags: u64,
+    pub anchors: *mut *mut c_char,
+    pub anchors_nr: usize,
 }
 
-// Re-export the xdiff API used by src/diff.c, forwarding to the Rust implementation
-// provided by the `diff` crate.  This allows the C code to link against the Rust
-// diff algorithm instead of the former C implementation.
+#[repr(C)]
+pub struct xdemitcb_t {
+    pub priv_: *mut c_void,
+    pub out_hunk: Option<
+        unsafe extern "C" fn(
+            *mut c_void,
+            c_long,
+            c_long,
+            c_long,
+            c_long,
+            *const c_char,
+            c_long,
+        ) -> c_int,
+    >,
+    pub out_line: Option<unsafe extern "C" fn(*mut c_void, *mut mmbuffer_t, c_int) -> c_int>,
+}
 
-#[no_mangle]
-pub unsafe extern "C" fn xdl_diff(
+#[repr(C)]
+pub struct xdemitconf_t {
+    pub ctxlen: c_long,
+    pub interhunkctxlen: c_long,
+    pub flags: u64,
+    pub find_func: Option<
+        unsafe extern "C" fn(*const c_char, c_long, *mut c_char, c_long, *mut c_void) -> c_long,
+    >,
+    pub find_func_priv: *mut c_void,
+    pub hunk_func:
+        Option<unsafe extern "C" fn(c_long, c_long, c_long, c_long, *mut c_void) -> c_int>,
+}
+
+/// Perform a diff between two mmfiles and emit the result through callbacks.
+/// Mirrors a subset of the libxdiff `xdl_diff` interface.
+///
+/// # Safety
+/// All pointers must be valid for the duration of the call.
+pub unsafe fn xdl_diff(
     mf1: *const mmfile_t,
     mf2: *const mmfile_t,
-    xpp: *const xpparam_t,
-    xecfg: *const xdemitconf_t,
+    _xpp: *const xpparam_t,
+    _xecfg: *const xdemitconf_t,
     ecb: *mut xdemitcb_t,
 ) -> c_int {
-    rs_xdl_diff(mf1, mf2, xpp, xecfg, ecb)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn xdiff_out_unified(
-    priv_: *mut c_void,
-    mb: *mut mmbuffer_t,
-    nbuf: c_int,
-) -> c_int {
-    rs_xdiff_out_unified(priv_, mb, nbuf)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn xdiff_out_indices(
-    start_a: c_long,
-    count_a: c_long,
-    start_b: c_long,
-    count_b: c_long,
-    priv_: *mut c_void,
-) -> c_int {
-    rs_xdiff_out_indices(start_a, count_a, start_b, count_b, priv_)
+    let res = std::panic::catch_unwind(|| {
+        if mf1.is_null() || mf2.is_null() || ecb.is_null() {
+            return -1;
+        }
+        let a_slice = std::slice::from_raw_parts((*mf1).ptr as *const u8, (*mf1).size as usize);
+        let b_slice = std::slice::from_raw_parts((*mf2).ptr as *const u8, (*mf2).size as usize);
+        let a_str = std::str::from_utf8(a_slice).unwrap_or("");
+        let b_str = std::str::from_utf8(b_slice).unwrap_or("");
+        let a_lines: Vec<&str> = if a_str.is_empty() {
+            Vec::new()
+        } else {
+            a_str.split_inclusive('\n').collect()
+        };
+        let b_lines: Vec<&str> = if b_str.is_empty() {
+            Vec::new()
+        } else {
+            b_str.split_inclusive('\n').collect()
+        };
+        let diff = TextDiff::configure().diff_slices(&a_lines, &b_lines);
+        if let Some(cb) = (*ecb).out_line {
+            for change in diff.iter_all_changes() {
+                let val = change.value();
+                let mut line = Vec::with_capacity(val.len() + 1);
+                line.push(match change.tag() {
+                    ChangeTag::Equal => b' ',
+                    ChangeTag::Insert => b'+',
+                    ChangeTag::Delete => b'-',
+                });
+                line.extend_from_slice(val.as_bytes());
+                let mut buf = mmbuffer_t {
+                    ptr: line.as_mut_ptr() as *mut c_char,
+                    size: line.len() as c_long,
+                };
+                cb((*ecb).priv_, &mut buf, 1);
+            }
+        }
+        0
+    });
+    res.unwrap_or(-1)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::ffi::CString;
-    use std::ffi::CStr;
-    use std::time::Instant;
+    use std::ffi::c_void;
 
-    #[test]
-    fn diff_basic() {
-        let a = "a\nb\n";
-        let b = "a\nc\n";
-        let diff = diff_internal(a, b);
-        assert!(diff.contains("-b"));
-        assert!(diff.contains("+c"));
+    unsafe extern "C" fn collect(priv_: *mut c_void, buf: *mut mmbuffer_t, _nr: c_int) -> c_int {
+        let out = &mut *(priv_ as *mut String);
+        let slice = std::slice::from_raw_parts((*buf).ptr as *const u8, (*buf).size as usize);
+        out.push_str(std::str::from_utf8(slice).unwrap());
+        0
     }
 
     #[test]
-    fn ffi_roundtrip() {
-        let a = CString::new("one\ntwo\n").unwrap();
-        let b = CString::new("one\nthree\n").unwrap();
-        let ptr = vim_xdiff_diff(a.as_ptr(), b.as_ptr());
-        let diff = unsafe { CStr::from_ptr(ptr).to_str().unwrap().to_owned() };
-        vim_xdiff_free(ptr);
-        assert!(diff.contains("-two"));
-        assert!(diff.contains("+three"));
-    }
-
-    #[test]
-    fn performance_small() {
-        let a = "a\n".repeat(1000);
-        let mut b = "a\n".repeat(999);
-        b.push_str("b\n");
-        let start = Instant::now();
-        let _ = diff_internal(&a, &b);
-        // ensure the diff runs quickly for small inputs
-        assert!(start.elapsed().as_millis() < 200);
-    }
-
-    #[test]
-    fn performance_large() {
-        let a = "a\n".repeat(10_000);
-        let mut b = "a\n".repeat(9_999);
-        b.push_str("b\n");
-        let start = Instant::now();
-        let _ = diff_internal(&a, &b);
-        // basic performance check for larger inputs
-        assert!(start.elapsed().as_millis() < 2000);
+    fn diff_simple() {
+        let a = b"a\nb\n";
+        let b = b"a\nc\n";
+        let mf1 = mmfile_t {
+            ptr: a.as_ptr() as *const c_char,
+            size: a.len() as c_long,
+        };
+        let mf2 = mmfile_t {
+            ptr: b.as_ptr() as *const c_char,
+            size: b.len() as c_long,
+        };
+        let mut output = String::new();
+        let mut ecb = xdemitcb_t {
+            priv_: &mut output as *mut _ as *mut c_void,
+            out_hunk: None,
+            out_line: Some(collect),
+        };
+        let res = unsafe { xdl_diff(&mf1, &mf2, std::ptr::null(), std::ptr::null(), &mut ecb) };
+        assert_eq!(res, 0);
+        assert!(output.contains("-b"));
+        assert!(output.contains("+c"));
     }
 }
